requirements.txt
dynaconf==3.1.12
python-dotenv==1.0.0
SQLAlchemy==2.0.9
pydantic==1.10.7
fastapi==0.95.1
loguru==0.5.3
humps==0.2.2
uvicorn==0.21.1
